refactor: create a separate component for task display (filtered)
